:package: change import paths
This might be a wrong GitMoji.
The gist of it, is that I changed the import paths for most of the big Python modules (files). This was ChatGPT's solution.
Now main is run by `py -m app.main`, which enables *pytest* to be run from same directory level with `pytest -v`.