:coffin: :bug: :pencil2: remove parameter `me` from `FunctionAssignment.update( )`
Why was it spelled `me` and not `name` in the first place??

This implementation, currently, doesn't support updating the functions name as it would mean we would have to delete the old function, both from the bidirectional mapping (field `function_to_nurses` in `TaskBoard`) and all `Nurse` objects which has that `FunctionAssignment`.
Alas, it probably isn't a problem as if there is an update like that, then I would just remake the entire `TaskBoard`. It is neither big enough, nor updated frequently enough for this to be a problem.