:adhesive_bandage: ♻️ Refactor -  `extract_task( )` needs `cell ` param to be `str`
The function that was previously called `valid_task( )` is now simply just `regex_filtering( )`, because I had to incur `str_and_non_empty( )` earlier, more precisely in `arbejdsplan_daily_tasks_lists( )`, for the cell to be `str` for `extract_task`.
ALSO: doesn't check whether the cell is not NaN, cause at this stage it is already established that it is both a `str` and non-empty.